Timer: Introduce Optimal Timer Frequency for Maximum Efficiency
After going through several articles on the matter and making experiments with
different frequencies, and assuming the calculations for the rate of kernel update
are based on "1000/Effective HZ = Rate of Kernel Update in ms", where "Effective HZ"
is taken from "NR_CPUS * CONFIG_HZ", this frequency was chosen and has yielded the
best results in terms of efficiency without compromising performance or causing
issues.
Although all of this matter still needs better documentation and is often cause of
debate, the beta testing results on this specific platform (8 Core - Kirin960)
at least have proven to be effective.

Signed-off-by: nutcasev15 <tab.andsupp@gmail.com>
Signed-off-by: JBolho <hugo.ss.android@gmail.com>
Tested-by: JBolho <hugo.ss.android@gmail.com>